casacore_lattices

Struct SubLattice

pub struct SubLattice<'a, T: LatticeElement, L: Lattice<T>> { /* private fields */ }

A read-only view of a lattice restricted by a region.

Corresponds to the C++ SubLattice<T>. The sub-lattice presents a view whose coordinate system is relative to the region’s bounding box. Pixels outside the region are returned as the default value for type T.

Examples

use casacore_lattices::*;
use ndarray::{ArrayD, IxDyn};

let data = ArrayD::from_shape_fn(IxDyn(&[10, 10]), |idx| (idx[0] * 10 + idx[1]) as f64);
let lat = ArrayLattice::new(data);
let region = LCBox::new(vec![2, 3], vec![5, 7], vec![10, 10]);
let sub = SubLattice::new(&lat, Box::new(region));

assert_eq!(sub.shape(), &[4, 5]); // bounding box shape
assert_eq!(sub.get_at(&[0, 0]).unwrap(), 23.0); // lat[2, 3]

Implementations

impl<'a, T: LatticeElement, L: Lattice<T>> SubLattice<'a, T, L>

pub fn new(parent: &'a L, region: Box<dyn LCRegion>) -> Self

Creates a sub-lattice view restricted by the given region.

pub fn region(&self) -> &dyn LCRegion

Returns the region defining this sub-lattice.